Update admin controller

+++ src/main/java/com/manpibrook/backend_api/repository/CategoryRepository.java
@@ -5,18 +5,15 @@ import java.util.Optional;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
-
-import com.manpibrook.backend_api.dto.response.CategoryResponse;
 import com.manpibrook.backend_api.entity.Category;
 @Repository
-public interface CategoryRepository extends JpaRepository<Category,Long>{
-	// create a pruduct need check category is exsist
-	Optional<Category> findByCategoryId(Long id);
-	// get all category order by name for category controller
-	List<Category> findAllByOrderByName();
-	// Delete category by id
-	void deleteByCategoryId(Long id);
-	
-	boolean existsByCategoryId(Long id);
-	
+public interface CategoryRepository extends JpaRepository<Category, Integer> {
+
+    Optional<Category> findByCategoryId(Integer id);
+
+    List<Category> findAllByOrderByName();
+
+    void deleteByCategoryId(Integer id);
+
+    boolean existsByCategoryId(Integer id);
 }
